+++ app/src/main/java/com/hongmyeoun/goldcalc/view/search/content/Content.kt
@@ -1,0 +1,64 @@
+package com.hongmyeoun.goldcalc.view.search.content
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyListState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavHostController
+import com.hongmyeoun.goldcalc.ui.theme.ImageBG
+import com.hongmyeoun.goldcalc.view.common.LoadingScreen
+import com.hongmyeoun.goldcalc.viewModel.search.SearchVM
+
+@Composable
+fun SearchContent(
+    paddingValues: PaddingValues,
+    viewModel: SearchVM,
+    navController: NavHostController,
+    scrollState: LazyListState,
+    onTap: () -> Unit
+) {
+    val isLoading by viewModel.isLoading.collectAsState()
+    val errorMessage by viewModel.errorMessage.collectAsState()
+    val characterList by viewModel.characterList.collectAsState()
+
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(paddingValues)
+            .pointerInput(Unit) {
+                detectTapGestures(
+                    onTap = { onTap() }
+                )
+            }
+            .background(ImageBG)
+            .padding(top = 8.dp),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Top
+    ) {
+
+        when {
+            isLoading -> {
+                LoadingScreen()
+            }
+
+            errorMessage != null -> {
+                Error(errorMessage)
+            }
+
+            else -> {
+                SearchResults(characterList, navController, scrollState)
+            }
+        }
+    }
+}
